Split economy service into ledger package for easier testing
